python/AnythingConvert/requirements.txt
# 图片处理
Pillow>=10.0.0
opencv-python>=4.8.0
cairosvg>=2.7.0

# 视频音频处理  
ffmpeg-python>=0.2.0

# 文档处理
python-docx>=0.8.11
pypandoc>=1.15
PyPDF2>=3.0.1
reportlab>=4.0.0
docx2txt>=0.8
odfpy>=1.4.0
chardet>=5.0.0  # 用于编码检测

# 命令行界面
click>=8.1.0
typer>=0.9.0
rich>=13.0.0

# Web界面
gradio>=4.0.0

# 工具库
tqdm>=4.66.0